server/requirements.txt
marshmallow==3.20.1
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy
typing_extensions==4.8.0
flask-restx
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.1.3
Flask==2.3.3
flask-restx
gunicorn==21.2.0
idna==3.4
importlib-resources==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
pluggy==1.3.0
pymongo==4.5.0
pyrsistent==0.19.3
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.0
